fix(#388): cast days to ::int in make_interval for defensive typing
PR #388 round-3 WARNING: psycopg might bind Python int as bigint,
and `make_interval(days => bigint)` doesn't exist. Cast explicitly
to `int` — matches the pattern elsewhere in the codebase (e.g.
app/api/instruments.py:404). Tests still pass.

+++ app/api/portfolio.py
@@ -717,7 +717,7 @@ def get_rolling_pnl(
                     SELECT close FROM price_daily
                     WHERE instrument_id = p.instrument_id
                       AND close IS NOT NULL
-                      AND price_date <= curr.price_date - make_interval(days => %(days)s)
+                      AND price_date <= curr.price_date - make_interval(days => %(days)s::int)
                     ORDER BY price_date DESC
                     LIMIT 1
                 ) prior ON TRUE
